fix(modal unit tests): spell check/move utils
"fix #1615"
