+++ projects/cashmere/src/lib/datepicker/calendar/calendar.component.spec.ts
@@ -3,8 +3,8 @@ import {Component} from '@angular/core';
 import {waitForAsync, ComponentFixture, inject, TestBed} from '@angular/core/testing';
 import {By} from '@angular/platform-browser';
 import {DatepickerModule} from '../datepicker.module';
-import {JAN, FEB, DEC, NOV, JUL} from '../utils/month-constants';
-import {dispatchFakeEvent, dispatchMouseEvent, dispatchEvent} from '../utils/dispatch-events';
+import {JAN, FEB, DEC, NOV, JUL} from '../../utils/month-constants';
+import {dispatchFakeEvent, dispatchMouseEvent, dispatchEvent} from '../../utils/dispatch-events';
 import {HcNativeDateModule} from '../datetime/datetime.module';
 import {DateAdapter} from '../datetime/date-adapter';
 import {CalendarComponent} from './calendar.component';
