+++ projects/cashmere/src/lib/datepicker/datepicker-input/datepicker-input.directive.ts
@@ -11,7 +11,7 @@ import {
 } from '@angular/forms';
 import {createMissingDateImplError} from '../datetime/datepicker-errors';
 import {DatepickerComponent} from '../datepicker.component';
-import {coerceBooleanProperty} from '../utils/boolean-property';
+import {coerceBooleanProperty} from '../../utils/boolean-property';
 import {Subscription} from 'rxjs';
 import {D, HC_DATE_FORMATS, HcDateFormats} from '../datetime/date-formats';
 import {DateAdapter} from '../datetime/date-adapter';
