+++ projects/cashmere/src/lib/datepicker/datetime/native-date-adapter.spec.ts
@@ -1,7 +1,7 @@
 import {Platform} from '@angular/cdk/platform';
 import {LOCALE_ID} from '@angular/core';
 import {waitForAsync, inject, TestBed} from '@angular/core/testing';
-import {JAN, DEC, FEB, MAR} from '../utils/month-constants';
+import {JAN, DEC, FEB, MAR} from '../../utils/month-constants';
 import {DateAdapter, HC_DATE_LOCALE} from './date-adapter';
 import {NativeDateAdapter} from './native-date-adapter';
 import {NativeDateModule} from './datetime.module';
