+++ projects/cashmere/src/lib/datepicker/multi-year-view/multi-year-view.component.spec.ts
@@ -4,8 +4,8 @@ import {waitForAsync, ComponentFixture, TestBed} from '@angular/core/testing';
 import {By} from '@angular/platform-browser';
 import {HcNativeDateModule} from '../datetime/datetime.module';
 import {yearsPerPage, yearsPerRow, MultiYearViewComponent} from './multi-year-view.component';
-import {JAN} from '../utils/month-constants';
-import {dispatchFakeEvent, dispatchEvent} from '../utils/dispatch-events';
+import {JAN} from '../../utils/month-constants';
+import {dispatchFakeEvent, dispatchEvent} from '../../utils/dispatch-events';
 import {CalendarBodyComponent} from '../calendar-body/calendar-body.component';
 
 /* tslint:disable */
