--- projects/cashmere/src/lib/modal/utils/boolean-property.ts
@@ -1,4 +1,0 @@
-/** Coerces a data-bound value (typically a string) to a boolean. */
-export function coerceBooleanProperty(value: any): boolean {
-    return value != null && `${value}` !== 'false';
-}
