--- projects/cashmere/src/lib/modal/utils/dispatch-events.ts
@@ -1,25 +1,0 @@
-/**
- * @license
- * Copyright Health Catalyst All Rights Reserved.
- *
- * Use of this source code is governed by an Apache-2.0 license that can be
- * found in the LICENSE file at https://raw.githubusercontent.com/HealthCatalyst/Fabric.Cashmere/dev/LICENSE
- */
-
-import {createFakeEvent, createMouseEvent} from './event-objects';
-
-/** Utility to dispatch any event on a Node. */
-export function dispatchEvent(node: Node | Window, event: Event): Event {
-    node.dispatchEvent(event);
-    return event;
-}
-
-/** Shorthand to dispatch a fake event on a specified node. */
-export function dispatchFakeEvent(node: Node | Window, type: string, canBubble?: boolean): Event {
-    return dispatchEvent(node, createFakeEvent(type, canBubble));
-}
-
-/** Shorthand to dispatch a mouse event on the specified coordinates. */
-export function dispatchMouseEvent(node: Node, type: string, x = 0, y = 0, event = createMouseEvent(type, x, y)): MouseEvent {
-    return dispatchEvent(node, event) as MouseEvent;
-}
